+++ core/src/main/java/com/novoda/noplayer/internal/drm/provision/ProvisionExecutorImpl.java
@@ -1,4 +1,4 @@
-package com.novoda.noplayer.drm.provision;
+package com.novoda.noplayer.internal.drm.provision;
 
 import com.novoda.noplayer.drm.ModularDrmProvisionRequest;
